+++ DeltaDNA/DDNALog.h
@@ -33,11 +33,20 @@
     #define DDNA_DEBUG NO
 #endif
 
-#define DDNALogDebug(frmt, ...) do{ if(DDNA_DEBUG) [DDNALog log:DDNA_LOG_FLAG_DEBUG \
-                                                           file:__FILE__ \
-                                                       function:sel_getName(_cmd) \
-                                                           line:__LINE__ \
-                                                         format:(frmt), ##__VA_ARGS__]; } while(0)
+typedef NS_OPTIONS(NSUInteger, DDNALogLevel) {
+    DDNALogLevelNone    = 0,
+    DDNALogLevelError   = 1 << 0,
+    DDNALogLevelWarn    = 1 << 1,
+    DDNALogLevelInfo    = 1 << 2,
+    DDNALogLevelDebug   = 1 << 3,
+    DDNALogLevelVerbose = 1 << 4
+};
+
+#define DDNALogDebug(frmt, ...) do{ [DDNALog log:DDNA_LOG_FLAG_DEBUG \
+                                            file:__FILE__ \
+                                        function:sel_getName(_cmd) \
+                                            line:__LINE__ \
+                                          format:(frmt), ##__VA_ARGS__]; } while(0)
 
 #define DDNALogWarn(frmt, ...) do{ [DDNALog log:DDNA_LOG_FLAG_WARN \
                                            file:__FILE__ \
@@ -47,6 +56,8 @@
 
 
 @interface DDNALog : NSObject
+
++ (void)setLogLevel:(DDNALogLevel)logLevel;
 
 + (void)log:(int)flag
        file:(const char *)file