+++ Framework/DeltaDNA.h
@@ -29,6 +29,7 @@ FOUNDATION_EXPORT const unsigned char DeltaDNAVersionString[];
 #define _DELTADNA_
 
 #import <DeltaDNA/DDNASDK.h>
+#import <DeltaDNA/DDNALog.h>
 #import <DeltaDNA/DDNASettings.h>
 #import <DeltaDNA/DDNAParams.h>
 #import <DeltaDNA/DDNAEvent.h>